pyproject.toml
[tool.poetry]
name = "savanamed"
version = "0.1.0"
description = "Savanamed programming task"
authors = ["Pedro José Piquero Plaza <gowikel@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
inject = "^4.2.0"
pdftotext = "^2.1.4"
python-magic = "^0.4.18"
faker = "^4.1.1"
pymongo = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
ipython = "^7.16.1"
flask-shell-ipython = "^0.4.1"
flake8 = "^3.8.3"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
